Split code into packages
