+++ Xamarin/Social-tap/Code/MainActivityController.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using Android.App;
 using Android.OS;
+using Android.Widget;
+using RestSharp;
 using Socialtap.Model;
 
-namespace Socialtap
+namespace Socialtap.Code
 {
     public static class MainActivityController
     {
@@ -21,14 +24,19 @@ namespace Socialtap
             }
         }
 
-        public static Boolean AddBarReview(String barName = "Nenurodyta", int beverageLevel = 0,
+        public static Boolean AddBarReview(string barName = "Nenurodyta", int beverageLevel = 0,
                                     String comment = "Nėra", int rating = 0)
         {
-            // pereiti per listą ieškant, ar nėra jau tokio baro
-            // yra:
-            // ...
-            // nėra:
-            // ...
+            var client = new RestClient("http://localhost:5050");
+            var request = new RestRequest("resource/{id}", Method.POST);
+
+            request.AddJsonBody(new BarData(barName, beverageLevel, comment, rating));
+
+            // execute the request
+            IRestResponse response = client.Execute(request);
+            var content = response.Content; // raw content as string
+            Toast.MakeText(Application.Context,
+                           content, ToastLength.Short).Show();
             return false;
         }
