+++ Xamarin/Social-tap/Code/Model/IBarData.cs
