+++ Xamarin/Social-tap/Social-tap.csproj
@@ -8,7 +8,7 @@
     <OutputType>Library</OutputType>
     <RootNamespace>Socialtap</RootNamespace>
     <AssemblyName>Social-tap</AssemblyName>
-    <TargetFrameworkVersion>v7.0</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v8.0</TargetFrameworkVersion>
     <AndroidApplication>True</AndroidApplication>
     <AndroidResgenFile>Resources\Resource.designer.cs</AndroidResgenFile>
     <AndroidResgenClass>Resource</AndroidResgenClass>
@@ -42,7 +42,7 @@
     <Reference Include="System.Xml" />
     <Reference Include="System.Core" />
     <Reference Include="Mono.Android" />
-      <Reference Include="Xamarin.Android.Support.Annotations">
+    <Reference Include="Xamarin.Android.Support.Annotations">
       <HintPath>..\packages\Xamarin.Android.Support.Annotations.25.3.1\lib\MonoAndroid70\Xamarin.Android.Support.Annotations.dll</HintPath>
     </Reference>
     <Reference Include="Xamarin.Android.Support.Compat">
@@ -102,16 +102,26 @@
     <Reference Include="RestSharp">
       <HintPath>..\packages\RestSharp.105.2.3\lib\MonoAndroid10\RestSharp.dll</HintPath>
     </Reference>
+    <Reference Include="Xamarin.Android.Support.v7.CardView">
+      <HintPath>..\packages\Xamarin.Android.Support.v7.CardView.25.3.1\lib\MonoAndroid70\Xamarin.Android.Support.v7.CardView.dll</HintPath>
+    </Reference>
+    <Reference Include="Xamarin.Android.Support.Constraint.Layout.Solver">
+      <HintPath>..\packages\Xamarin.Android.Support.Constraint.Layout.Solver.1.0.2.2\lib\MonoAndroid70\Xamarin.Android.Support.Constraint.Layout.Solver.dll</HintPath>
+    </Reference>
+    <Reference Include="Xamarin.Android.Support.Constraint.Layout">
+      <HintPath>..\packages\Xamarin.Android.Support.Constraint.Layout.1.0.2.2\lib\MonoAndroid70\Xamarin.Android.Support.Constraint.Layout.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="MainActivity.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Model\IBarData.cs" />
-    <Compile Include="Model\BarData.cs" />
-    <Compile Include="MainActivityController.cs" />
     <Compile Include="Resources\Resource.designer.cs" />
-    <Compile Include="ReviewFragment.cs" />
-    <Compile Include="BarListFragment.cs" />
+    <Compile Include="Code\BarListRecyclerView\BarListAdapter.cs" />
+    <Compile Include="Code\Model\BarData.cs" />
+    <Compile Include="Code\Model\IBarData.cs" />
+    <Compile Include="Code\MainActivity.cs" />
+    <Compile Include="Code\MainActivityController.cs" />
+    <Compile Include="Code\Fragments\BarListFragment.cs" />
+    <Compile Include="Code\Fragments\ReviewFragment.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
@@ -140,11 +150,16 @@
     <AndroidResource Include="Resources\drawable\ic_rate_review.xml" />
     <AndroidResource Include="Resources\drawable\ic_camera.xml" />
     <AndroidResource Include="Resources\layout\fragment_review.axml" />
+    <AndroidResource Include="Resources\layout\fragment_bar_list.axml" />
+    <AndroidResource Include="Resources\layout\fragment_bar_list_item.axml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Resources\drawable\" />
-    <Folder Include="Model\" />
     <Folder Include="Resources\values-v21\" />
+    <Folder Include="Code\" />
+    <Folder Include="Code\BarListRecyclerView\" />
+    <Folder Include="Code\Model\" />
+    <Folder Include="Code\Fragments\" />
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <Import Project="..\packages\Xamarin.Build.Download.0.4.3\build\Xamarin.Build.Download.targets" Condition="Exists('..\packages\Xamarin.Build.Download.0.4.3\build\Xamarin.Build.Download.targets')" />
@@ -167,4 +182,5 @@
   <Import Project="..\packages\Xamarin.GooglePlayServices.InstantApps.42.1021.1\build\MonoAndroid70\Xamarin.GooglePlayServices.InstantApps.targets" Condition="Exists('..\packages\Xamarin.GooglePlayServices.InstantApps.42.1021.1\build\MonoAndroid70\Xamarin.GooglePlayServices.InstantApps.targets')" />
   <Import Project="..\packages\Xamarin.GooglePlayServices.Location.42.1021.1\build\MonoAndroid70\Xamarin.GooglePlayServices.Location.targets" Condition="Exists('..\packages\Xamarin.GooglePlayServices.Location.42.1021.1\build\MonoAndroid70\Xamarin.GooglePlayServices.Location.targets')" />
   <Import Project="..\packages\Xamarin.GooglePlayServices.Maps.42.1021.1\build\MonoAndroid70\Xamarin.GooglePlayServices.Maps.targets" Condition="Exists('..\packages\Xamarin.GooglePlayServices.Maps.42.1021.1\build\MonoAndroid70\Xamarin.GooglePlayServices.Maps.targets')" />
+  <Import Project="..\packages\Xamarin.Android.Support.v7.CardView.25.3.1\build\MonoAndroid70\Xamarin.Android.Support.v7.CardView.targets" Condition="Exists('..\packages\Xamarin.Android.Support.v7.CardView.25.3.1\build\MonoAndroid70\Xamarin.Android.Support.v7.CardView.targets')" />
 </Project>
