+++ Xamarin/Social-tap/packages.config
@@ -4,6 +4,8 @@
   <package id="Xamarin.Android.Support.Animated.Vector.Drawable" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Annotations" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Compat" version="25.3.1" targetFramework="monoandroid71" />
+  <package id="Xamarin.Android.Support.Constraint.Layout" version="1.0.2.2" targetFramework="monoandroid70" />
+  <package id="Xamarin.Android.Support.Constraint.Layout.Solver" version="1.0.2.2" targetFramework="monoandroid70" />
   <package id="Xamarin.Android.Support.Core.UI" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Core.Utils" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Design" version="25.3.1" targetFramework="monoandroid71" />
@@ -12,6 +14,7 @@
   <package id="Xamarin.Android.Support.Transition" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.v4" version="25.3.1" targetFramework="monoandroid70" />
   <package id="Xamarin.Android.Support.v7.AppCompat" version="25.3.1" targetFramework="monoandroid71" />
+  <package id="Xamarin.Android.Support.v7.CardView" version="25.3.1" targetFramework="monoandroid70" />
   <package id="Xamarin.Android.Support.v7.RecyclerView" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Vector.Drawable" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Build.Download" version="0.4.3" targetFramework="monoandroid71" />
